requirements.txt
# Audio recording
sounddevice==0.5.1
soundfile==0.13.1
numpy==2.2.3

# APIs
openai==1.68.2
anthropic==0.49.0

# Desktop integration
pystray==0.19.5
Pillow==11.1.0
pynput==1.7.7
pyperclip==1.9.0

# HTTP client (for Docker mode)
requests==2.32.3

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1
